+++ src/hooks/useUserAffiliations/useUserAffiliations.test.js
@@ -1,11 +1,14 @@
-import { renderHook } from '@testing-library/react-hooks';
+import { renderHook } from '@folio/jest-config-stripes/testing-library/react-hooks';
 import {
   QueryClient,
   QueryClientProvider,
 } from 'react-query';
 import orderBy from 'lodash/orderBy';
 
-import { useOkapiKy } from '@folio/stripes/core';
+import {
+  useOkapiKy,
+  useStripes,
+} from '@folio/stripes/core';
 
 import affiliations from '../../../test/jest/fixtures/affiliations';
 import consortia from '../../../test/jest/fixtures/consortia';
@@ -14,15 +17,14 @@ import {
   CONSORTIA_USER_TENANTS_API,
   MAX_RECORDS,
 } from '../../constants';
-import useConsortium from '../useConsortium';
 import useUserAffiliations from './useUserAffiliations';
 
 jest.mock('@folio/stripes/core', () => ({
   ...jest.requireActual('@folio/stripes/core'),
   useNamespace: jest.fn(() => ['test']),
   useOkapiKy: jest.fn(),
+  useStripes: jest.fn(),
 }));
-jest.mock('../useConsortium', () => jest.fn());
 
 const queryClient = new QueryClient();
 
@@ -40,7 +42,7 @@ const response = {
 
 const consortium = {
   ...consortia[0],
-  centralTenant: 'mobius',
+  centralTenantId: 'mobius',
 };
 
 describe('useUserAffiliations', () => {
@@ -60,8 +62,12 @@ describe('useUserAffiliations', () => {
 
   beforeEach(() => {
     mockGet.mockClear();
-    useConsortium.mockClear().mockReturnValue({ consortium });
     useOkapiKy.mockClear().mockReturnValue(kyMock);
+    useStripes.mockClear().mockReturnValue({
+      user: {
+        user: { consortium },
+      }
+    });
   });
 
   it('should fetch user\'s consortium affiliations by user\'s id when there is consortium', async () => {
@@ -79,7 +85,7 @@ describe('useUserAffiliations', () => {
   });
 
   it('should not fetch user\'s consortium affiliations by user\'s id when there is not consortium', async () => {
-    useConsortium.mockClear().mockReturnValue({ consortium: {} });
+    useStripes.mockClear().mockReturnValue({ a: 1 });
 
     const userId = 'usedId';
     const { result, waitFor } = renderHook(() => useUserAffiliations({ userId }), { wrapper });