+++ src/settings/TransferAccountsSettings.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import { screen } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
+import { screen } from '@folio/jest-config-stripes/testing-library/react';
+import userEvent from '@folio/jest-config-stripes/testing-library/user-event';
 import { ControlledVocab } from '@folio/stripes/smart-components';
 
 import '__mock__/stripesCore.mock';