+++ CHANGELOG.md
@@ -29,6 +29,9 @@
 * "ui-users.settings.customfields.view" permission insufficient to view custom fields on user settings. Refs UIU-2863.
 * Align affiliation assignment with stripes-core updates (switch active affiliation). Refs UIU-2855.
 * Create Jest/RTL test for UserEdit.js. Refs UIU-2426
+* Also support `circulation` `14.0`. Refs UIU-2858.
+* Import `@testing-library` deps from `jest-config-stripes`. Refs UIU-2866.
+* Also support `request-storage` `6.0`. Refs UIU-2875.
 
 ## [9.0.0](https://github.com/folio-org/ui-users/tree/v9.0.0) (2023-02-20)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v8.1.0...v9.0.0)